Bump linters
