+++ tests/_async_compat/__init__.py
@@ -14,8 +14,7 @@
 # limitations under the License.
 
 
-from functools import wraps as _wraps
-
+from ._util import wrap_async
 from .mark_decorator import (
     async_fixture,
     AsyncTestDecorators,
@@ -35,13 +34,3 @@ __all__ = [
     "mark_sync_test",
     "wrap_async",
 ]
-
-
-def wrap_async(func):
-    @_wraps(func)
-    async def wrapper(*args, **kwargs):  # noqa: RUF029
-        # [noqa] the hole point of this wrapper is to turn a sync function into
-        # an async one for testing purposes
-        return func(*args, **kwargs)
-
-    return wrapper
