+++ tests/integration/examples/test_config_trust_example.py
@@ -16,7 +16,7 @@
 
 import pytest
 
-from . import DriverSetupExample
+from ._base import DriverSetupExample
 
 
 # isort: off
